+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 <!--Copyright Laserfiche.
 Licensed under the MIT License. See LICENSE in the project root for license information.-->
+## 18.1.0
+
+### Breaking Changes
+ - Updated services to use the Repository API v2 instead of Repository API v1
+ - Please pin `lf-ui-components-services` to version `18.0.2` if your project is still using Repository API v1
+
 ## 18.0.2
 
 ### Chore & Maintenance